Добавляет URL в href, выносит отдельно data, JSdoc
