+++ js/utilities.js
@@ -1,7 +1,7 @@
-// Выбирает случайное число из массива:
+/** Выбирает случайное число из массива */
 const getRandomArrayElement = (elements) => elements[Math.floor(Math.random() * elements.length)];
 
-// Выбирает положительное число из диапазона:
+/** Выбирает положительное число из диапазона */
 const getRandomPositiveInteger = (min, max) => {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -9,7 +9,7 @@ const getRandomPositiveInteger = (min, max) => {
   return Math.floor(result);
 };
 
-// Генерирует ID:
+/** Генерирует ID */
 const createRandomIdFromRangeGenerator = (min, max) => {
   const previousValues = [];
 
@@ -26,22 +26,22 @@ const createRandomIdFromRangeGenerator = (min, max) => {
   };
 };
 
-// Закрывает окно
+/** Закрывает окно */
 const closesModal = (item) => {
   document.body.classList.remove('modal-open');
   item.classList.add('hidden');
 };
 
-// Открывает окно
+/** Открывает окно */
 const opensModal = (item) => {
   document.body.classList.add('modal-open');
   item.classList.remove('hidden');
 };
 
-// Проверяет массив на наличие повторяющихся элементов
+/** Проверяет массив на наличие повторяющихся элементов */
 const checksDuplicateElements = (arr) => arr.length === new Set(arr).size;
 
-// Приводит строку к нижнему регистру и удаляет концевые пробелы
+/** Приводит строку к нижнему регистру и удаляет концевые пробелы */
 const normalizeString = (str) => str.trim().split(' ').filter((string) => Boolean(string.length));
 
 export { getRandomArrayElement, getRandomPositiveInteger, createRandomIdFromRangeGenerator, closesModal, opensModal, checksDuplicateElements, normalizeString };
